Use roboelectric for testing for performance reasons
