+++ app/build.gradle
@@ -19,8 +19,8 @@ android {
         applicationId "de.tum.in.tumcampus"
         minSdkVersion 14
         targetSdkVersion 27
-        versionCode 99999
-        versionName "1.5.1-dev"
+        versionCode 516
+        versionName "1.5.0"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
     }
     lintOptions {
@@ -42,7 +42,11 @@ android {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
-
+    testOptions {
+        unitTests {
+            includeAndroidResources = true
+        }
+    }
 
     publishNonDefault true
 }
@@ -90,6 +94,11 @@ dependencies {
     androidTestImplementation 'org.hamcrest:hamcrest-library:1.3'
     androidTestImplementation "com.android.support.test.espresso:espresso-core:$espressoVersion"
     androidTestImplementation "com.android.support.test.espresso:espresso-contrib:$espressoVersion"
+
+    //Used for testing
+    testImplementation 'junit:junit:4.12'
+    testImplementation 'org.mockito:mockito-core:1.10.19'
+    testImplementation 'org.robolectric:robolectric:3.5.1'
 }
 
 apply plugin: 'com.google.gms.google-services'
