--- app/src/androidTest/java/de/tum/in/tumcampusapp/activities/CalendarActivityTest.java
@@ -1,21 +1,0 @@
-package de.tum.in.tumcampusapp.activities;
-
-import android.support.test.filters.LargeTest;
-import android.support.test.rule.ActivityTestRule;
-import android.support.test.runner.AndroidJUnit4;
-
-import org.junit.Rule;
-import org.junit.runner.RunWith;
-
-@RunWith(AndroidJUnit4.class)
-@LargeTest
-public class CalendarActivityTest extends BaseActivityTest {
-
-    @Rule
-    public ActivityTestRule<CalendarActivity> mActivityRule = new ActivityTestRule<>(CalendarActivity.class);
-
-    @Override
-    public void mainComponentDisplayedTest() {
-        // TODO
-    }
-}
