--- app/src/androidTest/java/de/tum/in/tumcampusapp/activities/MainActivityTest.java
@@ -1,33 +1,0 @@
-package de.tum.in.tumcampusapp.activities;
-
-import android.support.test.filters.LargeTest;
-import android.support.test.rule.ActivityTestRule;
-import android.support.test.runner.AndroidJUnit4;
-
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import de.tum.in.tumcampusapp.R;
-
-import static android.support.test.espresso.Espresso.onView;
-import static android.support.test.espresso.assertion.ViewAssertions.matches;
-import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
-import static android.support.test.espresso.matcher.ViewMatchers.withText;
-
-@RunWith(AndroidJUnit4.class)
-@LargeTest
-public class MainActivityTest extends BaseActivityTest {
-
-    @Rule
-    public ActivityTestRule<MainActivity> mActivityRule = new ActivityTestRule<>(MainActivity.class);
-
-    @Test
-    @Override
-    public void mainComponentDisplayedTest() {
-        idIsDisplayed(R.id.cards_view);
-
-        onView(withText(R.string.swipe_instruction))
-                .check(matches(isDisplayed()));
-    }
-}
