+++ app/src/test/java/de/tum/in/tumcampusapp/activities/CurriculaActivityTest.java
@@ -1,0 +1,13 @@
+package de.tum.in.tumcampusapp.activities;
+
+import org.junit.runner.RunWith;
+import org.robolectric.RobolectricTestRunner;
+
+@RunWith(RobolectricTestRunner.class)
+public class CurriculaActivityTest extends BaseActivityTest {
+
+    @Override
+    public void mainComponentDisplayedTest() {
+       // idIsDisplayed(R.id.activity_curricula_list_view);
+    }
+}
