+++ app/src/test/java/de/tum/in/tumcampusapp/activities/MainActivityTest.java
@@ -1,0 +1,18 @@
+package de.tum.in.tumcampusapp.activities;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.robolectric.RobolectricTestRunner;
+
+@RunWith(RobolectricTestRunner.class)
+public class MainActivityTest extends BaseActivityTest {
+
+
+    @Test
+    @Override
+    public void mainComponentDisplayedTest() {
+        //idIsDisplayed(R.id.cards_view);
+
+        //onView(withText(R.string.swipe_instruction)).check(matches(isDisplayed()));
+    }
+}
